Deprecate `str::from_utf8_lossy`
Use `String::from_utf8_lossy` instead

[breaking-change]
