+++ src/test/run-make/unicode-input/multiple_files.rs
@@ -8,7 +8,7 @@
 // option. This file may not be copied, modified, or distributed
 // except according to those terms.
 
-use std::{char, os, str};
+use std::{char, os};
 use std::io::{File, Command};
 use std::rand::{task_rng, Rng};
 
@@ -61,7 +61,7 @@ fn main() {
                                           main_file.as_str()
                                                    .unwrap()).as_slice())
                              .output().unwrap();
-        let err = str::from_utf8_lossy(result.error.as_slice());
+        let err = String::from_utf8_lossy(result.error.as_slice());
 
         // positive test so that this test will be updated when the
         // compiler changes.
